Delete ratchet tree code
